requirements.txt
annotated-types==0.6.0
anthropic==0.19.1
anyio==4.3.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.8.0
blessed==1.20.0
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
command-not-found==0.3
contourpy==1.2.1
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.12.1
dbus-python==1.2.18
distro==1.8.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.0
fastlogging==1.0.0
filelock==3.13.1
flake8==7.1.1
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httplib2==0.20.2
httpx==0.25.2
huggingface-hub==0.21.4
idna==3.4
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
keybert==0.8.4
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
markdownify==0.11.6
MarkupSafe==2.1.3
matplotlib==3.8.4
mccabe==0.7.0
mdurl==0.1.2
mistletoe==1.3.0
more-itertools==8.10.0
mplfinance==0.12.10b0
mpmath==1.3.0
mypy-extensions==1.0.0
netifaces==0.11.0
netlify-py==0.1.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
ollama==0.1.7
openai==1.14.2
oscrypto==1.3.0
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pdfminer.six==20231228
peewee==3.16.3
pillow==10.2.0
platformdirs==4.3.6
playwright==1.42.0
pluggy==1.4.0
prompt-toolkit==3.0.40
psutil==5.9.6
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pyee==11.0.1
pyflakes==3.2.0
Pygments==2.17.2
PyGObject==3.42.1
pyHanko==0.23.1
pyhanko-certvalidator==0.26.3
PyJWT==2.3.0
pyparsing==2.4.7
pypdf==4.1.0
pypng==0.20220715.0
pytest==7.4.2
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-apt==2.4.0+ubuntu3
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-editor==1.0.4
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
questionary==1.10.0
readchar==4.0.5
referencing==0.34.0
regex==2023.10.3
reportlab==4.0.9
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
schedule==1.2.2
scikit-learn==1.4.1.post1
scipy==1.12.0
SecretStorage==3.3.1
sentence-transformers==2.6.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlmodel==0.0.16
svglib==1.5.1
sympy==1.12
systemd-python==234
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.4.0
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
torch==2.2.1
tqdm==4.66.2
transformers==4.39.1
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritools==4.0.2
urllib3==1.26.7
uv==0.1.24
wadllib==1.3.6
wcwidth==0.2.8
webencodings==0.5.1
Werkzeug==3.0.1
xhtml2pdf==0.2.15
yaspin==2.5.0
zipp==1.0.0
